refactor: move credentials logic to client component, and handle redirects for other providers server-side
