+++ keep-ui/app/(signin)/signin/page.tsx
@@ -1,5 +1,6 @@
-"use client";
-import SignInForm from "./SignInForm";
+import { signIn, providerMap } from "@/auth";
+import { CredentialsSignInForm } from "./CredentialsSignInForm";
+import { SignInLoader } from "./ui/loader";
 
 export type SignInPageProps = {
   params: {
@@ -10,6 +11,34 @@ export type SignInPageProps = {
   };
 };
 
-export default function SignInPage({ params, searchParams }: SignInPageProps) {
-  return <SignInForm params={params} searchParams={searchParams} />;
+export default async function SignInPage({
+  params,
+  searchParams,
+}: SignInPageProps) {
+  const redirectTo = searchParams.callbackUrl ?? "/";
+  if (providerMap.get("auth0")) {
+    console.log("Signing in with auth0 provider");
+    if (params?.amt) {
+      signIn("auth0", { redirectTo }, { acr_values: `amt:${params.amt}` });
+    } else {
+      signIn("auth0", { redirectTo });
+    }
+  } else if (providerMap.get("keycloak")) {
+    console.log("Signing in with keycloak provider");
+    signIn("keycloak", { redirectTo });
+  } else if (providerMap.get("microsoft-entra-id")) {
+    console.log("Signing in with Azure AD provider");
+    signIn("microsoft-entra-id", { redirectTo });
+  } else if (
+    providerMap.get("credentials") &&
+    providerMap.get("credentials")?.name == "NoAuth"
+  ) {
+    signIn("credentials", { redirectTo });
+  }
+  if (providerMap.get("credentials")) {
+    return (
+      <CredentialsSignInForm params={params} searchParams={searchParams} />
+    );
+  }
+  return <SignInLoader text="Redirecting to authentication..." />;
 }
